feat: Linked Todo and User together, including auth

+++ backend/src/controllers/todoController.ts
@@ -1,23 +1,27 @@
 import { Todo, TodoModel } from "@/models/todo";
 
-export async function indexTodos() {
-    const todos = await TodoModel.find().sort({ createdAt: -1 });
+export async function indexTodos(userId: string) {
+    const todos = await TodoModel.find({ userId }).sort({ createdAt: -1 });
     return todos;
 }
 
-export async function createTodo({
-    title,
-    task,
-}: {
-    title: string;
-    task?: string;
-}) {
-    console.log(title);
-    if (!title) {
+export async function createTodo(
+    userId: string,
+    {
+        title,
+        task,
+    }: {
+        userId: string;
+        title: string;
+        task?: string;
+    }
+) {
+    if (!title || !userId) {
         throw new Error("Invalid body");
     }
 
     const newTodo = new TodoModel({
+        userId,
         title: title,
         task: task || "",
     });
@@ -26,18 +30,25 @@ export async function createTodo({
     return newTodo;
 }
 
-export async function showTodo(id: string) {
-    const todo = await TodoModel.findById(id);
+export async function showTodo(userId: string, id: string) {
+    const todo = await TodoModel.findOne({ _id: id, userId });
 
     if (!todo) {
         throw new Error("Invalid Id");
     }
 
+    if (todo.userId?.toString() !== userId) {
+        throw new Error("Not authorized");
+    }
+
     return todo;
 }
 
-export async function editTodo(id: string, todo: Todo) {
-    const newTodo = await TodoModel.findOneAndUpdate({ _id: id }, { ...todo });
+export async function editTodo(userId: string, id: string, todo: Todo) {
+    const newTodo = await TodoModel.findOneAndUpdate(
+        { _id: id, userId },
+        { ...todo }
+    );
 
     if (!newTodo) {
         throw new Error("Invalid Id");
@@ -46,10 +57,10 @@ export async function editTodo(id: string, todo: Todo) {
     return newTodo;
 }
 
-export async function destroyTodo(id: string) {
-	const result = await TodoModel.findOneAndDelete({ _id: id });
+export async function destroyTodo(userId: string, id: string) {
+    const result = await TodoModel.findOneAndDelete({ _id: id, userId });
 
-	if (!result) {
+    if (!result) {
         throw new Error("Invalid Id");
-	}
+    }
 }
